backend/pyproject.toml
[tool.poetry]
name = "lut-generator"
version = "0.1.0"
description = "Custom LUT Generator with LLM and Image Analysis"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
python-multipart = "^0.0.6"
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"
python-dotenv = "^1.0.0"
llama-cpp-python = "^0.2.6"
opencv-python-headless = "^4.8.1.78"
pillow = "^10.0.1"
numpy = "^1.24.3"
scipy = "^1.11.3"
aiofiles = "^23.2.1"
structlog = "^23.1.0"
colorlog = "^6.7.0"
tenacity = "^8.2.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.9.1"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.5.1"
pre-commit = "^3.4.0"
httpx = "^0.24.1"
pytest-mock = "^3.11.1"
faker = "^19.6.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = [
    "cv2.*",
    "llama_cpp.*"
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --cov=app --cov-report=term-missing"
testpaths = [
    "tests"
]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["app"]
omit = [
    "tests/*",
    "**/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
    "raise ImportError",
]

[tool.poetry.scripts]
start = "app.main:start"
dev = "app.main:dev"